requirements.txt
langchain==0.2.16
langchain-community==0.2.16
langchain-groq==0.1.9
chromadb==0.5.5
sentence-transformers==3.0.1
pymupdf==1.24.9
streamlit==1.38.0
python-dotenv==1.0.1